requirements/production.txt
Django
dj-database-url
psycopg2
gunicorn
honcho
requests
celery
flower
redis
django-pipeline
djangorestframework
markdown
